# docs/swagger.yaml
definitions:
  controller._ResponseAllClassesList:
    properties:
      code:
        $ref: '#/definitions/controller.ResponseCodeType'
      data:
        description: '`json:"data,omitempty"` //omitempty 该字段为空时忽略'
        items:
          $ref: '#/definitions/models.ResponseClassBrief'
        type: array
      msg: {}
    type: object
...
  controller._ResponseClassDetailList:
    properties:
      code:
        $ref: '#/definitions/controller.ResponseCodeType'
      data:
        allOf:
        - $ref: '#/definitions/models.ResponseClassDetail'
        description: '`json:"data,omitempty"` //omitempty 该字段为空时忽略'
      msg: {}
    type: object
  controller._ResponseCreateClass:
    properties:
      code:
        $ref: '#/definitions/controller.ResponseCodeType'
      data:
        allOf:
        - $ref: '#/definitions/models.ResponseClassBrief'
        description: '`json:"data,omitempty"` //omitempty 该字段为空时忽略'
      msg: {}
    type: object
  controller._ResponseDeleteClass:
    properties:
      code:
        $ref: '#/definitions/controller.ResponseCodeType'
      data:
        items:
          type: integer
        type: array
      msg: {}
    type: object
  controller._ResponseNoDataArea:
    properties:
      code:
        $ref: '#/definitions/controller.ResponseCodeType'
      msg: {}
    type: object
...
  controller.ResponseCodeType:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 1011
    - 1012
    - 1013
    - 1014
    - 1015
    - 1016
    - 1017
    - 1018
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeParameterInvalid
    - CodeUserNotExist
    - CodeUserExist
    - CodePasswordError
    - CodeServerBusy
    - CodeTokenInvaild
    - CodeTokenEmpty
    - CodeNeedReLogin
    - CodeCommunityIdInvalid
    - CodeVerifyCodeInvaild
    - CodeArticleTitleExisted
    - CodeArticleNotExisted
    - CodeHaveArticleInClass
    - CodeTagExisted
    - CodeTagNotExisted
    - CodeClassNotExisted
    - CodeClassNameExisted
    - CodeCreateCommentFailed
...
  models.ClassCreateFormParams:
    properties:
      desc:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  models.CommentCreateFormParams:
    properties:
      article_id:
        description: 被评论的文章ID
        example: "0"
        type: string
      content:
        description: 评论内容
        type: string
      email:
        description: 评论人的邮箱 - 不强制要求
        type: string
      parent_comment_id:
        description: RootCommentId   uint   `json:"root_comment_id,string"`   // 根评论的ID
        example: "0"
        type: string
      status:
        description: 评论状态：0-审核中 1-审核通过 2-审核未通过
        type: integer
      type:
        description: 评论种类 1- 表示文章评论
        type: integer
      user_name:
        description: 评论人名字 - 强制要求
        type: string
    type: object
  models.CommentDeleteFormParams:
    properties:
      ids:
        description: id列表
        items:
          type: integer
        type: array
    type: object
...
  models.DeleteMultiTagParams:
    properties:
      ids:
        description: 待删除的TagID列表
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
...
  models.ResponseClassBrief:
    properties:
      atc_count:
        description: 该分类下包含的文章数量
        type: integer
      class_id:
        description: 类别ID - 由应用层生成
        example: "0"
        type: string
      name:
        description: 类别名称
        type: string
    type: object
  models.ResponseClassDetail:
    properties:
      atc_count:
        description: 该分类下包含的文章数量
        type: integer
      class_id:
        description: 类别ID - 由应用层生成
        example: "0"
        type: string
      created_at:
        description: 创建时间
        type: string
      desc:
        description: 类别的描述信息
        type: string
      name:
        description: 类别名称
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
...
  models.UpdateClassParams:
    properties:
      class_id:
        example: "0"
        type: string
      desc:
        type: string
      name:
        type: string
    required:
    - class_id
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: NgBlog Go博客项目 API 接口文档
  title: NgBlog
  version: "1.0"
paths:
...
  /class:
    post:
      consumes:
      - application/json
      description: 通过该接口可以创建类别
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      - description: 创建类别的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.ClassCreateFormParams'
      produces:
      - application/json
      responses:
        "200":
          description: code=1000表示成功其余失败
          schema:
            $ref: '#/definitions/controller._ResponseCreateClass'
      security:
      - ApiKeyAuth: []
      summary: 创建类别(Class)的接口
      tags:
      - 类别相关接口
    put:
      consumes:
      - application/json
      description: 通过该接口可以更新类别
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      - description: 更新类别的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.UpdateClassParams'
      produces:
      - application/json
      responses:
        "200":
          description: code=1000表示成功其余失败
          schema:
            $ref: '#/definitions/controller._ResponseNoDataArea'
      security:
      - ApiKeyAuth: []
      summary: 更新类别(Class)的接口
      tags:
      - 类别相关接口
  /class/:id:
    delete:
      consumes:
      - application/json
      description: 通过该接口可以删除类别
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code=1000表示成功
          schema:
            $ref: '#/definitions/controller._ResponseNoDataArea'
      security:
      - ApiKeyAuth: []
      summary: 删除类别(Class)的接口
      tags:
      - 类别相关接口
    get:
      consumes:
      - application/json
      description: 通过该接口可以获得指定ID的类别
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code字段为1000表示执行成功，其余表示出错
          schema:
            $ref: '#/definitions/controller._ResponseClassDetailList'
      security:
      - ApiKeyAuth: []
      summary: 根据类别ID查询类的信息（详细信息）
      tags:
      - 类别相关接口
  /classes:
    delete:
      consumes:
      - application/json
      description: 通过该接口可以删除多个类别
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      - description: 待删除的class ID列表
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.DeleteMultiTagParams'
      produces:
      - application/json
      responses:
        "200":
          description: code=1000成功，,返回data域中不为空代表对应的class删除失败
          schema:
            $ref: '#/definitions/controller._ResponseDeleteClass'
      security:
      - ApiKeyAuth: []
      summary: 删除多个类别(Class)的接口
      tags:
      - 类别相关接口
    get:
      consumes:
      - application/json
      description: 通过该接口可以获得当前的所有文章
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseAllClassesList'
      security:
      - ApiKeyAuth: []
      summary: 获取所有类别（简略信息）的接口
      tags:
      - 类别相关接口
  /comment:
    post:
      consumes:
      - application/json
      description: 通过该接口可以创建评论
      parameters:
      - description: 创建评论的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.CommentCreateFormParams'
      produces:
      - application/json
      responses:
        "200":
          description: code=1000表示成功其余失败
          schema:
            $ref: '#/definitions/controller._ResponseNoDataArea'
      security:
      - ApiKeyAuth: []
      summary: 创建评论Comment的接口
      tags:
      - 评论相关接口
  /comment/:id:
    delete:
      consumes:
      - application/json
      description: 通过该接口可以删除单条评论
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除的评论的ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code=1000表示成功其余失败
          schema:
            $ref: '#/definitions/controller._ResponseNoDataArea'
      security:
      - ApiKeyAuth: []
      summary: 删除单条评论Comment的接口
      tags:
      - 评论相关接口
  /comments:
    delete:
      consumes:
      - application/json
      description: 通过该接口可以修改评论的状态
      parameters:
      - description: Bearer token(jwt)
        in: header
        name: Authorization
        required: true
        type: string
      - description: 删除的评论的ID
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.CommentDeleteFormParams'
      produces:
      - application/json
      responses:
        "200":
          description: code=1000表示成功其余失败
          schema:
            $ref: '#/definitions/controller._ResponseNoDataArea'
      security:
      - ApiKeyAuth: []
      summary: 修改指定评论的状态的接口
      tags:
      - 评论相关接口
    get:
      consumes:
      - application/json
      description: 通过该接口可以分页获取评论
      parameters:
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 评论类型 1-文章
        in: query
        name: type
        type: integer
      - description: 对应的文章id
        in: query
        name: atc_id
        type: string
      - description: '状态: 1-审核中 2-审核通过 3-审核未通过'
        in: query
        name: status
        type: integer
      - description: 页号
        in: query
        name: page_idx
        type: integer
      - description: 每页大小
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code=1000表示成功其余失败
          schema:
            $ref: '#/definitions/controller._ResponseNoDataArea'
      security:
      - ApiKeyAuth: []
      summary: 分页获取评论的接口
      tags:
      - 评论相关接口
...
swagger: "2.0"